change port 6xxx to 8xxx because of local services on Macos change postgres column type to decimal(20)  , because no support of unsigned bigint on Postgres

+++ test/postgres/create_tab_client_activity.sql
@@ -1,6 +1,6 @@
 CREATE TABLE client_activity (
   clientid bigint NOT NULL,
-  ticketid bigint NOT NULL,
+  ticketid decimal(20) NOT NULL,
   state int NOT NULL,
   modified timestamptz NULL DEFAULT NULL,
   border smallint   DEFAULT '0',
